Move util to `testonly`
